Reorganised Header Files

+++ Source/Incursion_CPP/Private/W_HUD_Ammo.cpp
@@ -8,7 +8,7 @@ void UW_HUD_Ammo::Initialise_Implementation()
 	// OVERRIDE IN BP AND SET TextBlockAmmo BEFORE PARENT INITIALISE
 }
 
-void UW_HUD_Ammo::SetAmmoAmount(int MaxAmmo, int CurrentAmmo)
+void UW_HUD_Ammo::SetAmmoAmount(int8 MaxAmmo, int8 CurrentAmmo)
 {
 	FText AmmoText = FText::Format(FText::FromString(TEXT("{0} / {1}")), FText::AsNumber(CurrentAmmo), FText::AsNumber(MaxAmmo));
 	TextBlockAmmo->SetText(AmmoText);
